define structs in include/rclc/types.h
